+++ src/crewai/rag/__init__.py
@@ -1,0 +1,1 @@
+"""RAG (Retrieval-Augmented Generation) infrastructure for CrewAI."""
