+++ src/crewai/rag/embeddings/__init__.py
@@ -1,0 +1,1 @@
+"""Embedding components for RAG infrastructure."""
