+++ src/crewai/rag/storage/__init__.py
@@ -1,0 +1,1 @@
+"""Storage components for RAG infrastructure."""
